disassemble video decode module in videoanalysis app
