+++ smartcity/videoanalysis/videoanalysisapp/common/include/video_analysis_params.h
@@ -40,22 +40,31 @@
 using hiai::ImageData;
 using hiai::IMAGEFORMAT;
 
+// video type
+enum VideoType {
+  kH264,
+  kH265,
+  kInvalidTpye
+};
+
 struct VideoImageInfoT {
   std::string channel_id;
   uint32_t frame_id;
   std::string channel_name;
   bool is_finished;
+  VideoType video_type;
 
   VideoImageInfoT& operator=(VideoImageInfoT& value) {
     channel_id = value.channel_id;
     frame_id = value.frame_id;
     channel_name = value.channel_name;
     is_finished = value.is_finished;
+    video_type = value.video_type;
     return *this;
   }
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, VideoImageInfoT& data) {
   ar(data.channel_id, data.frame_id, data.channel_name, data.is_finished);
 }
@@ -65,7 +74,7 @@ struct VideoImageParaT {
   hiai::ImageData<u_int8_t> img;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, VideoImageParaT& data) {
   ar(data.video_image_info, data.img);
 }
@@ -75,7 +84,7 @@ struct ObjectInfoT {
   float score;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, ObjectInfoT& data) {
   ar(data.object_id, data.score);
 }
@@ -85,7 +94,7 @@ struct ObjectImageParaT {
   hiai::ImageData<u_int8_t> img;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, ObjectImageParaT& data) {
   ar(data.object_info, data.img);
 }
@@ -95,7 +104,7 @@ struct VideoDetectionImageParaT {
   std::vector<ObjectImageParaT> obj_imgs;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, VideoDetectionImageParaT& data) {
   ar(data.image, data.obj_imgs);
 }
@@ -105,7 +114,7 @@ struct BatchCroppedImageParaT {
   std::vector<ObjectImageParaT> obj_imgs;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, BatchCroppedImageParaT& data) {
   ar(data.video_image_info, data.obj_imgs);
 }
@@ -118,12 +127,12 @@ enum CarInferenceType {
 struct CarInfoT {
   std::string object_id;
   uint32_t label;
-  CarInferenceType attribute_name;  // attribute name:cartype or carcolor
+  CarInferenceType attribute_name; // attribute name:cartype or carcolor
   std::string inference_result;
   float confidence;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, CarInfoT& data) {
   ar(data.object_id, data.label, data.attribute_name, data.inference_result,
      data.confidence);
@@ -134,7 +143,7 @@ struct BatchCarInfoT {
   std::vector<CarInfoT> car_infos;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, BatchCarInfoT& data) {
   ar(data.video_image_info, data.car_infos);
 }
@@ -145,7 +154,7 @@ struct OutputT {
   std::shared_ptr<u_int8_t> data;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, OutputT& data) {
   ar(data.size);
   ar(data.name);
@@ -163,20 +172,18 @@ struct DetectionEngineTransT {
   VideoImageParaT video_image;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, DetectionEngineTransT& data) {
   ar(data.status, data.msg, data.output_datas, data.video_image);
 }
 
 struct PedestrianInfoT {
   std::string object_id;
-  std::string
-      attribute_name;  // property name:cartype or carcolor or pedestrian
-  std::map<string, float>
-      pedestrian_attribute_map;  // string:label_name,float:confidence
+  std::string attribute_name; // property name:cartype or carcolor or pedestrian
+  std::map<string, float> pedestrian_attribute_map; // string:label_name,float:confidence
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, PedestrianInfoT& data) {
   ar(data.object_id, data.attribute_name, data.pedestrian_attribute_map);
 }
@@ -186,7 +193,7 @@ struct BatchPedestrianInfoT {
   std::vector<PedestrianInfoT> pedestrian_info;
 };
 
-template <class Archive>
+template<class Archive>
 void serialize(Archive& ar, BatchPedestrianInfoT& data) {
   ar(data.video_image_info, data.pedestrian_info);
 }
