+++ smartcity/videoanalysis/videoanalysisapp/object_detection/object_detection.cpp
@@ -37,6 +37,15 @@
 #include <sstream>
 #include "ascenddk/ascend_ezdvpp/dvpp_process.h"
 
+#include "hiaiengine/ai_model_manager.h"
+#include "hiaiengine/ai_tensor.h"
+#include "hiaiengine/ai_types.h"
+#include "hiaiengine/api.h"
+#include "hiaiengine/data_type.h"
+#include "hiaiengine/data_type_reg.h"
+#include "hiaiengine/engine.h"
+#include "video_analysis_params.h"
+
 using ascend::utils::DvppCropOrResizePara;
 using ascend::utils::DvppOutput;
 using ascend::utils::DvppProcess;
@@ -49,20 +58,65 @@ namespace {
 const uint32_t kInputPort = 0;
 const uint32_t kOutputPort = 0;
 
+const int kHandleSuccessful = 0; // the process handled successfully
+
+const int kVideoFormatLength = 5; // video format string length
+
+const int kImageDataQueueSize = 20; // the queue default size
+
+const int kWaitTime = 100000; // wait 100ms
+
+const string kVideoTypeH264 = "h264"; // video type h264
+const string kVideoTypeH265 = "h265"; // video type h265
+
+const string kImageFormatNv12 = "nv12"; // image format nv12
+
+const string kStrChannelId1 = "channel1"; // channle id 1 string
+const string kStrChannelId2 = "channel2"; // channle id 2 string
+
+const int kIntChannelId1 = 1; // channel id 1 integer
+const int kIntChannelId2 = 2; // channel id 2 integer
+
+// The width of image in vpc interface need 128-byte alignment
+const int kVpcWidthAlign = 128;
+
+// The height of image in vpc interface need 16-byte alignment
+const int kVpcHeightAlign = 16;
+
+// standard: 4096 * 4096 * 4 = 67108864 (64M)
+const int kAllowedMaxImageMemory = 67108864;
+
+const int kMaxRetryNumber = 100; // the maximum number of retry
+
+const int kCompareEqual = 0; // string compare equal
+
 // ssd input image width and height.
 const uint32_t kInputWidth = 512;
 const uint32_t kInputHeight = 512;
 
-const int kDvppProcSuccess = 0;     // call dvpp success return.
-const int kSleepMicroSecs = 20000;  // waiting time after queue is full.
+const int kDvppProcSuccess = 0; // call dvpp success return.
+const int kSleepMicroSecs = 20000; // waiting time after queue is full.
 
 const string kModelPath = "model_path";
-}  // namespace
+
+// the queue store yuv image data
+ThreadSafeQueue<shared_ptr<VideoImageParaT>> yuv_image_queue(
+    kImageDataQueueSize);
+
+uint32_t frame_id_1 = 0; // frame id used for channle1
+
+uint32_t frame_id_2 = 0; // frame id used for channle2
+}
 
 HIAI_REGISTER_DATA_TYPE("VideoImageInfoT", VideoImageInfoT);
 HIAI_REGISTER_DATA_TYPE("VideoImageParaT", VideoImageParaT);
 HIAI_REGISTER_DATA_TYPE("OutputT", OutputT);
 HIAI_REGISTER_DATA_TYPE("DetectionEngineTransT", DetectionEngineTransT);
+
+ObjectDetectionInferenceEngine::ObjectDetectionInferenceEngine() {
+  dvpp_api_channel1_ = nullptr;
+  dvpp_api_channel2_ = nullptr;
+}
 
 HIAI_StatusT ObjectDetectionInferenceEngine::Init(
     const hiai::AIConfig& config,
@@ -96,6 +150,226 @@ HIAI_StatusT ObjectDetectionInferenceEngine::Init(
 
   HIAI_ENGINE_LOG(HIAI_DEBUG_INFO, "[ODInferenceEngine] engine initialized!");
   return HIAI_OK;
+}
+
+void CallVpcGetYuvImage(FRAME* frame, void* hiai_data) {
+  if (frame == nullptr || hiai_data == nullptr) { // check input parameters
+    HIAI_ENGINE_LOG(
+        HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+        "The input data for function:CallVpcGetYuvImage is nullptr!");
+    return;
+  }
+
+  IDVPPAPI* dvpp_api = nullptr;
+  CreateDvppApi(dvpp_api);
+
+  if (dvpp_api == nullptr) { // check create dvpp api result
+    HIAI_ENGINE_LOG(
+        HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+        "Fail to call CreateDvppApi for vpc, the result is nullptr!");
+    return;
+  }
+
+  // constructing input image configuration
+  shared_ptr<VpcUserImageConfigure> image_configure(new VpcUserImageConfigure);
+  image_configure->widthStride = frame->width;
+  image_configure->heightStride = frame->height;
+
+  // check image format is nv12
+  if (strcmp(frame->image_format, kImageFormatNv12.c_str()) == kCompareEqual) {
+    image_configure->inputFormat = INPUT_YUV420_SEMI_PLANNER_UV;
+  } else { // check image format is nv21
+    image_configure->inputFormat = INPUT_YUV420_SEMI_PLANNER_VU;
+  }
+
+  image_configure->outputFormat = OUTPUT_YUV420SP_UV;
+  image_configure->isCompressData = true;
+  image_configure->compressDataConfigure.lumaHeadAddr = (long) (frame->buffer
+      + frame->offset_head_y);
+  image_configure->compressDataConfigure.chromaHeadAddr = (long) (frame->buffer
+      + frame->offset_head_c);
+  image_configure->compressDataConfigure.lumaHeadStride = frame->stride_head;
+  image_configure->compressDataConfigure.chromaHeadStride = frame->stride_head;
+  image_configure->compressDataConfigure.lumaPayloadAddr = (long) (frame->buffer
+      + frame->offset_payload_y);
+  image_configure->compressDataConfigure.chromaPayloadAddr = (long) (frame
+      ->buffer + frame->offset_payload_c);
+  image_configure->compressDataConfigure.lumaPayloadStride = frame
+      ->stride_payload;
+  image_configure->compressDataConfigure.chromaPayloadStride = frame
+      ->stride_payload;
+
+  shared_ptr<VpcUserRoiConfigure> roi_configure(new VpcUserRoiConfigure);
+  roi_configure->next = nullptr;
+
+  // constructing input roi configuration
+  VpcUserRoiInputConfigure *input_configure = &roi_configure->inputConfigure;
+  input_configure->cropArea.leftOffset = 0; // 0 means without crop
+  // dvpp limits rightOffset is odd
+  input_configure->cropArea.rightOffset =
+      frame->width % 2 == 0 ? frame->width - 1 : frame->width;
+  input_configure->cropArea.upOffset = 0; // 0 means without crop
+  // dvpp limits downOffset is odd
+  input_configure->cropArea.downOffset =
+      frame->height % 2 == 0 ? frame->height - 1 : frame->height;
+
+  int aligned_output_width = ALIGN_UP(frame->width, kVpcWidthAlign);
+  int aligned_output_height = ALIGN_UP(frame->height, kVpcHeightAlign);
+  int vpc_output_size = aligned_output_width * aligned_output_height
+      * DVPP_YUV420SP_SIZE_MOLECULE / DVPP_YUV420SP_SIZE_DENOMINATOR;
+
+  // check vpc output size is valid
+  if (vpc_output_size <= 0 || vpc_output_size > kAllowedMaxImageMemory) {
+    HIAI_ENGINE_LOG(
+        HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+        "The vpc_output_size:%d is invalid! value range: 1~67108864",
+        vpc_output_size);
+    return;
+  }
+
+  // construct vpc out data buffer
+  uint8_t *vpc_out_buffer = (uint8_t *) mmap(
+      0, vpc_output_size, PROT_READ | PROT_WRITE,
+      MAP_PRIVATE | MAP_ANONYMOUS | MAP_HUGETLB | API_MAP_VA32BIT, 0, 0);
+  if (vpc_out_buffer == MAP_FAILED) { // check new buffer result
+    HIAI_ENGINE_LOG("Failed to malloc memory in dvpp(new vpc).");
+    return;
+  }
+
+  // constructing output roi configuration
+  VpcUserRoiOutputConfigure *output_configure = &roi_configure->outputConfigure;
+  output_configure->addr = vpc_out_buffer;
+  output_configure->bufferSize = vpc_output_size;
+  output_configure->widthStride = frame->width;
+  output_configure->heightStride = frame->height;
+  output_configure->outputArea.leftOffset = 0; // 0 means without crop
+  // dvpp limits rightOffset is odd
+  output_configure->outputArea.rightOffset =
+      frame->width % 2 == 0 ? frame->width - 1 : frame->width;
+  output_configure->outputArea.upOffset = 0; // 0 means without crop
+  // dvpp limits downOffset is odd
+  output_configure->outputArea.downOffset =
+      frame->height % 2 == 0 ? frame->height - 1 : frame->height;
+
+  image_configure->roiConfigure = roi_configure.get();
+
+  dvppapi_ctl_msg dvpp_api_ctl_msg;
+  dvpp_api_ctl_msg.in = static_cast<void *>(image_configure.get());
+  dvpp_api_ctl_msg.in_size = sizeof(VpcUserImageConfigure);
+
+  // call vpc and check the result
+  if (DvppCtl(dvpp_api, DVPP_CTL_VPC_PROC, &dvpp_api_ctl_msg)
+      != kHandleSuccessful) {
+    HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                    "Fail to call dvppctl VPC!");
+    // free vpc_out_buffer memory
+    munmap(vpc_out_buffer, (unsigned) (ALIGN_UP(vpc_output_size, MAP_2M)));
+    DestroyDvppApi(dvpp_api);
+    return;
+  }
+
+  DestroyDvppApi(dvpp_api);
+
+  uint8_t* output_image_buffer = new (nothrow) uint8_t[vpc_output_size];
+  if (output_image_buffer == nullptr) { // check new result
+    HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                    "Fail to new data when handle vpc output!");
+    // free vpc_out_buffer memory
+    munmap(vpc_out_buffer, (unsigned) (ALIGN_UP(vpc_output_size, MAP_2M)));
+    return;
+  }
+
+  int memcpy_result = memcpy_s(output_image_buffer, vpc_output_size,
+                               vpc_out_buffer, vpc_output_size);
+  // free vpc_out_buffer memory
+  munmap(vpc_out_buffer, (unsigned) (ALIGN_UP(vpc_output_size, MAP_2M)));
+
+  if (memcpy_result != EOK) { // check memcpy_s result
+    HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                    "Fail to copy vpc output image buffer, memcpy_s result:%d",
+                    memcpy_result);
+    delete[] output_image_buffer;
+    return;
+  }
+
+  // send key frame data to next engine
+  HandleKeyFrameData(output_image_buffer, vpc_output_size, hiai_data, frame);
+  return;
+}
+
+void HandleKeyFrameData(uint8_t* image_data_buffer, uint32_t image_data_size,
+                        void* hiai_data, FRAME* frame) {
+  string channel_name = ((YuvImageFrameInfo*) (hiai_data))->channel_name;
+  string channel_id = ((YuvImageFrameInfo*) (hiai_data))->channel_id;
+  uint32_t frame_id = GetFrameId(channel_id);
+
+  // only send key frame to next engine, key frame id: 1,6,11,16...
+  if (!IsKeyFrame(frame_id)) {
+    delete[] image_data_buffer;
+    return;
+  }
+
+  HIAI_ENGINE_LOG("Get key frame, frame id:%d, channel_id:%s, channel_name:%s,"
+                  " frame->realWidth:%d, frame->realHeight:%d",
+                  frame_id, channel_id.c_str(), channel_name.c_str(),
+                  frame->realWidth, frame->realHeight);
+
+  //send yuv420sp data
+  VideoImageInfoT i_video_image_info;
+  i_video_image_info.channel_id = channel_id;
+  i_video_image_info.channel_name = channel_name;
+  i_video_image_info.frame_id = frame_id;
+  i_video_image_info.is_finished = false;
+
+  hiai::ImageData<unsigned char> image_data;
+  image_data.width = frame->realWidth;
+  image_data.height = frame->realHeight;
+  image_data.format = IMAGEFORMAT::YUV420SP;
+  image_data.size = image_data_size;
+
+  image_data.data.reset(image_data_buffer, default_delete<unsigned char[]>());
+  shared_ptr<VideoImageParaT> video_image_para = make_shared<VideoImageParaT>();
+  video_image_para->img = image_data;
+  video_image_para->video_image_info = i_video_image_info;
+
+  AddImage2QueueByChannel(video_image_para);
+}
+
+uint32_t GetFrameId(const string &channel_id) {
+  if (channel_id == kStrChannelId1) { // check input channel id is channel1
+    frame_id_1++;
+    return frame_id_1;
+  } else { // the input channel id is channel2
+    frame_id_2++;
+    return frame_id_2;
+  }
+}
+
+void AddImage2QueueByChannel(
+    const shared_ptr<VideoImageParaT>& video_image_para) {
+  bool add_image_success = false;
+  int count = 0; // count retry number
+
+  // add image data to queue, max retry time: 100
+  while (count < kMaxRetryNumber) {
+    count++;
+
+    if (yuv_image_queue.Push(video_image_para)) { // push image data to queue
+      add_image_success = true;
+      break;
+    } else { // queue is full
+      usleep(kWaitTime); // sleep 100 ms
+    }
+  }
+
+  if (!add_image_success) { // fail to send image data
+    HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                    "Fail to add image data to queue, channel_id:%s, "
+                    "channel_name:%s, frame_id:%d",
+                    video_image_para->video_image_info.channel_id.c_str(),
+                    video_image_para->video_image_info.channel_name.c_str(),
+                    video_image_para->video_image_info.frame_id);
+  }
 }
 
 HIAI_StatusT ObjectDetectionInferenceEngine::ImagePreProcess(
@@ -158,12 +432,12 @@ HIAI_StatusT ObjectDetectionInferenceEngine::PerformInference(
     shared_ptr<DetectionEngineTransT>& detection_trans,
     ImageData<u_int8_t>& input_img) {
   // init neural buffer.
-  shared_ptr<hiai::AINeuralNetworkBuffer> neural_buffer =
-      shared_ptr<hiai::AINeuralNetworkBuffer>(
-          new hiai::AINeuralNetworkBuffer(),
-          default_delete<hiai::AINeuralNetworkBuffer>());
+  shared_ptr<hiai::AINeuralNetworkBuffer> neural_buffer = shared_ptr<
+      hiai::AINeuralNetworkBuffer>(
+      new hiai::AINeuralNetworkBuffer(),
+      default_delete<hiai::AINeuralNetworkBuffer>());
 
-  neural_buffer->SetBuffer((void*)input_img.data.get(), input_img.size);
+  neural_buffer->SetBuffer((void*) input_img.data.get(), input_img.size);
 
   shared_ptr<hiai::IAITensor> input_tensor =
       static_pointer_cast<hiai::IAITensor>(neural_buffer);
@@ -182,8 +456,8 @@ HIAI_StatusT ObjectDetectionInferenceEngine::PerformInference(
   hiai::AIContext ai_context;
 
   // neural network inference.
-  ret =
-      ai_model_manager_->Process(ai_context, input_tensors, output_tensors, 0);
+  ret = ai_model_manager_->Process(ai_context, input_tensors, output_tensors,
+                                   0);
   if (ret != hiai::SUCCESS) {
     SendDetectionResult(detection_trans, false,
                         "[ODInferenceEngine] image inference failed!");
@@ -193,8 +467,8 @@ HIAI_StatusT ObjectDetectionInferenceEngine::PerformInference(
   // set trans_data
   detection_trans->status = true;
   for (uint32_t index = 0; index < output_tensors.size(); ++index) {
-    shared_ptr<hiai::AINeuralNetworkBuffer> result_tensor =
-        static_pointer_cast<hiai::AINeuralNetworkBuffer>(output_tensors[index]);
+    shared_ptr<hiai::AINeuralNetworkBuffer> result_tensor = static_pointer_cast<
+        hiai::AINeuralNetworkBuffer>(output_tensors[index]);
     OutputT out;
     out.size = result_tensor->GetSize();
     out.data = std::shared_ptr<uint8_t>(new uint8_t[out.size],
@@ -246,8 +520,112 @@ HIAI_StatusT ObjectDetectionInferenceEngine::SendDetectionResult(
   return HIAI_OK;
 }
 
+int ObjectDetectionInferenceEngine::GetIntChannelId(const string channel_id) {
+  if (channel_id == kStrChannelId1) { // check channel is channel1
+    return kIntChannelId1;
+  } else { // the channel is channel2
+    return kIntChannelId2;
+  }
+}
+
+bool ObjectDetectionInferenceEngine::ConvertH264ToHfbc(
+    const shared_ptr<VideoImageParaT>& video_image) {
+  IDVPPAPI* dvpp_api = nullptr;
+  if (video_image->video_image_info.channel_id == kStrChannelId1) {
+    dvpp_api = dvpp_api_channel1_;
+  } else {
+    dvpp_api = dvpp_api_channel2_;
+  }
+
+  if (dvpp_api == nullptr) {
+    CreateVdecApi(dvpp_api, 0);
+    if (dvpp_api == nullptr) { // check create dvpp api result
+      HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                      "[ODInferenceEngine] fail to create dvpp vdec api!");
+      return false;
+    }
+  }
+
+  vdec_in_msg vdec_msg;
+  vdec_msg.call_back = CallVpcGetYuvImage;
+  YuvImageFrameInfo yuv_image_frame_info;
+  yuv_image_frame_info.channel_name =
+      video_image->video_image_info.channel_name;
+  yuv_image_frame_info.channel_id = video_image->video_image_info.channel_id;
+  vdec_msg.channelId = GetIntChannelId(
+      video_image->video_image_info.channel_id);
+  vdec_msg.hiai_data = &yuv_image_frame_info;
+
+  int strcpy_result = EOK;
+  // check video type is h264 or h265
+  if (video_image->video_image_info.video_type == kH264) {
+    // video type: h264
+    strcpy_result = strcpy_s(vdec_msg.video_format, kVideoFormatLength,
+                             kVideoTypeH264.c_str());
+  } else { // the video type is h265
+    strcpy_result = strcpy_s(vdec_msg.video_format, kVideoFormatLength,
+                             kVideoTypeH265.c_str());
+  }
+
+  if (strcpy_result != EOK) {
+    // check strcpy result
+    HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                    "fail to call strcpy_s, result:%d, channel id:%s",
+                    strcpy_result,
+                    video_image->video_image_info.channel_id.c_str());
+    return false;
+  }
+
+  dvppapi_ctl_msg dvppApiCtlMsg;
+  dvppApiCtlMsg.in = (void*) ((&vdec_msg));
+  dvppApiCtlMsg.in_size = sizeof(vdec_in_msg);
+  // call vdec and check result
+  if (VdecCtl(dvpp_api, DVPP_CTL_VDEC_PROC, &dvppApiCtlMsg, 0)
+      != kHandleSuccessful) {
+    HIAI_ENGINE_LOG(HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+                    "Fail to call dvppctl process, channel id:%s",
+                    video_image->video_image_info.channel_id.c_str());
+    DestroyVdecApi(dvpp_api, 0);
+    return false;
+  }
+
+  return true;
+}
+
+void ObjectDetectionInferenceEngine::ObjectDetecitonInference(
+    shared_ptr<DetectionEngineTransT> detection_trans) {
+  while (!yuv_image_queue.empty()) {
+    shared_ptr<VideoImageParaT> video_image = yuv_image_queue.Pop();
+    if (video_image == nullptr) {
+      HIAI_ENGINE_LOG(
+          "[ODInferenceEngine] fail pop yuv image data from yuv_image_queue!");
+      continue;
+    }
+
+    detection_trans->video_image = *video_image;
+
+    // resize input image.
+    ImageData<u_int8_t> resized_img;
+    HIAI_StatusT dvpp_ret = ImagePreProcess(detection_trans->video_image.img,
+                                            resized_img);
+    if (dvpp_ret != HIAI_OK) {
+      // if preprocess error,send input image to the next engine.
+      SendDetectionResult(detection_trans, false,
+                          "[ODInferenceEngine] image preprocessed failure!");
+      return;
+    }
+
+    // inference object detection
+    if (PerformInference(detection_trans, resized_img) != HIAI_OK) {
+      HIAI_ENGINE_LOG(
+          HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+          "[ODInferenceEngine] fail to inference object detection!");
+    }
+  }
+}
+
 HIAI_IMPL_ENGINE_PROCESS("object_detection", ObjectDetectionInferenceEngine,
-                         INPUT_SIZE) {
+    INPUT_SIZE) {
   HIAI_ENGINE_LOG(HIAI_DEBUG_INFO, "[ODInferenceEngine] start process!");
 
   if (arg0 == nullptr) {
@@ -260,29 +638,36 @@ HIAI_IMPL_ENGINE_PROCESS("object_detection", ObjectDetectionInferenceEngine,
       static_pointer_cast<VideoImageParaT>(arg0);
 
   // init inference results tensor shared_ptr.
-  shared_ptr<DetectionEngineTransT> detection_trans =
-      make_shared<DetectionEngineTransT>();
+  shared_ptr<DetectionEngineTransT> detection_trans = make_shared<
+      DetectionEngineTransT>();
 
-  detection_trans->video_image = *video_image;
-  if (detection_trans->video_image.video_image_info.is_finished) {
+  detection_trans->video_image.video_image_info = video_image->video_image_info;
+  if (video_image->video_image_info.is_finished) {
     // input is finished.
     HIAI_ENGINE_LOG(HIAI_DEBUG_INFO,
                     "[ODInferenceEngine] input video finished!");
+
+    IDVPPAPI* dvpp_api = nullptr;
+    if (video_image->video_image_info.channel_id == kStrChannelId1) {
+      dvpp_api = dvpp_api_channel1_;
+    } else {
+      dvpp_api = dvpp_api_channel2_;
+    }
+
+    DestroyVdecApi(dvpp_api, 0);
+    ObjectDetecitonInference(detection_trans);
+
     return SendDetectionResult(detection_trans);
   }
 
-  // resize input image.
-  ImageData<u_int8_t> resized_img;
-  HIAI_StatusT dvpp_ret =
-      ImagePreProcess(detection_trans->video_image.img, resized_img);
-
-  if (dvpp_ret != HIAI_OK) {
-    // if preprocess error,send input image to the next engine.
-    SendDetectionResult(detection_trans, false,
-                        "[ODInferenceEngine] image preprocessed failure!");
+  if (!ConvertH264ToHfbc(video_image)) {
+    HIAI_ENGINE_LOG(
+        HIAI_ENGINE_RUN_ARGS_NOT_RIGHT,
+        "[ODInferenceEngine] fail to call dvpp vdec convert h264 to hfbc!");
     return HIAI_ERROR;
   }
 
-  // inference
-  return PerformInference(detection_trans, resized_img);
+  ObjectDetecitonInference(detection_trans);
+
+  return HIAI_OK;
 }
