+++ smartcity/videoanalysis/videoanalysisapp/object_detection/thread_safe_queue.h
